move applysourcestringrules to helpers
closes #389
