+++ src/helpers.ts
@@ -108,3 +108,64 @@ export function makeDirIfMissing(dir: string) {
     fse.mkdirSync(dir);
   }
 }
+
+const sourcesInfo: Record<number, string> = {};
+interface Categories {
+  sources: Record<
+    string, // a sourceTag. i.e. "adventures" or "deadorbit" or "zavala" or "crucible"
+    {
+      /**
+       * list of strings. if a sourceString contains one of these,
+       * it probably refers to this sourceTag
+       */
+      includes: string[];
+      /**
+       * list of strings. if a sourceString contains one of these,
+       * it doesn't refer to this sourceTag
+       */
+      excludes?: string[];
+      /** list of english item names or inventoryItem hashes */
+      items?: (string | number)[];
+      /** duplicate this category into another sourceTag */
+      alias?: string;
+      /**
+       * presentationNodes can contain a set of items (Collections).
+       * we'll find presentation nodes by name or hash,
+       * and include their children in this source
+       */
+      presentationNodes?: (string | number)[];
+      searchString?: string[];
+    }
+  >;
+  /** i don't really remember why this exists */
+  exceptions: string[][];
+}
+
+export function applySourceStringRules(
+  haystack: typeof sourcesInfo,
+  sourceStringRules: Categories['sources'][string]
+): number[] {
+  const { includes, excludes } = sourceStringRules;
+
+  return (
+    Object.entries(haystack)
+      // filter down to only search results that match these sourceStringRules
+      .filter(
+        ([, sourceString]) =>
+          // do inclusion strings match this sourceString?
+          includes?.filter((searchTerm) =>
+            sourceString.toLowerCase().includes(searchTerm.toLowerCase())
+          ).length &&
+          // not any excludes or not any exclude matches
+          !(
+            // do exclusion strings match this sourceString?
+            excludes?.filter((searchTerm) =>
+              sourceString.toLowerCase().includes(searchTerm.toLowerCase())
+            ).length
+          )
+      )
+      // keep the sourceHash and discard the sourceString.
+      // convert them back from object keys (strings) to numbers.
+      .map(([sourceHash]) => Number(sourceHash))
+  );
+}
